fix: separate the logic of the user schema and user validation
